Make collision test less resource-intensive
Instead of looping through the entire vector of solid tiles every time
one of them is removed for being too far away from the object doing the
testing, the iterator is now decremented. The end result of this is that
each tile is only tested once per object now, instead of multiple times
based on how many tiles are not in line with said object.

That's a lot of words for "changed j = -1 to j--", really.